refactor(code_editor): split 1,746-line file into 9 focused modules
Split monolithic code_editor.rs into:
- types.rs (83 lines): BracketPair, BracketMatch, IndentStyle, EditOp, EditorConfig
- navigation.rs (231 lines): Cursor and navigation methods
- selection.rs (101 lines): Selection methods
- editing.rs (378 lines): Insert, delete, undo/redo operations
- bracket.rs (141 lines): Bracket matching logic
- modes.rs (196 lines): Go-to-line and find functionality
- key_handling.rs (73 lines): Key event handling
- render.rs (361 lines): Rendering helpers and View trait impl
- tests.rs (51 lines): All 5 code_editor tests
- code_editor.rs (275 lines): Main struct, builder methods, exports

File reduced from 1,746 to 275 lines (modularized into 10 files).

Benefits:
- Single Responsibility: Each module has one clear purpose
- Easier navigation: Smaller, focused files
- Better testability: Tests separated into dedicated module
- Maintains backward compatibility: All public APIs unchanged
